+++ react/OrderForm.tsx
@@ -9,12 +9,12 @@ import React, {
 import { branch, renderComponent } from 'recompose'
 import { compose, graphql } from 'react-apollo'
 
-import OrderFormQuery from './graphql/orderForm.graphql'
+import { orderForm as OrderFormQuery } from 'vtex.checkout-resources/Queries'
 
 interface Context {
   loading: boolean
   orderForm: OrderForm | undefined
-  setOrderForm: (orderForm: OrderForm) => void
+  setOrderForm: (orderForm: Partial<OrderForm>) => void
 }
 
 interface OrderFormProviderProps {