+++ react/__mocks__/vtex.checkout-resources/Queries.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag'
+
+export const orderForm = gql`
+query MockQuery {
+  orderForm {
+    items
+  }
+}
+`